+++ src/main/java/kitchenpos/ordermaster/domain/OrderStatus.java
@@ -1,0 +1,9 @@
+package kitchenpos.ordermaster.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED;
+
+    public static OrderStatus initialOrderStatus() {
+        return WAITING;
+    }
+}
