+++ src/main/java/kitchenpos/ordermaster/domain/OrderType.java
@@ -1,0 +1,11 @@
+package kitchenpos.ordermaster.domain;
+
+import java.util.Objects;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN;
+
+    public static boolean isValid(OrderType orderType) {
+        return !Objects.isNull(orderType);
+    }
+}
